Reader ATmosphere: extract shared error-message helper
